+++ spec/scss_lint/linter/vendor_prefixes_spec.rb
@@ -347,4 +347,14 @@ describe SCSSLint::Linter::VendorPrefixes do
 
     it { should_not report_lint }
   end
+
+  context 'when property contains a list literal with an empty list' do
+    let(:scss) { <<-SCSS }
+      p {
+        content: 0 ();
+      }
+    SCSS
+
+    it { should_not report_lint }
+  end
 end